Refactor: 요청서 페이지 분리, 보여지는 요청서 파일 이동
