+++ client/front/lib/main/home/checkerrand/checkerrand.dart
@@ -6,10 +6,10 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_naver_map/flutter_naver_map.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
-import 'package:front/errands/fixerrand/fixerrandwidget/fixerrand.dart';
+import 'package:front/main/home/checkerrand/fixerrand/fixerrand.dart';
 import 'package:intl/intl.dart';
-import '../../mainshowerrand/mainshowerrand.dart';
-import '../home/home.dart';
+import 'showerrand/mainshowerrand.dart';
+import '../home.dart';
 import 'package:http/http.dart' as http;
 
 final storage = FlutterSecureStorage(); // 토큰 받기
