+++ client/front/lib/main/home/home.dart
@@ -6,7 +6,7 @@ import 'package:intl/intl.dart';
 import 'package:http/http.dart' as http;
 import 'dart:convert';
 import '../writeerrand/writeerrand.dart';
-import '../writeerrand/checkerrand.dart';
+import 'checkerrand/checkerrand.dart';
 import '../../status_page_requesting.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 final storage = FlutterSecureStorage();
